address code review comments
also move the resumable test cases to ObjectResumableOperationTest.cs
