+++ sdk/aliyun-oss-sdk.csproj
@@ -379,7 +379,7 @@
     <Compile Include="Transform\GetSymlinkResultDeserializer.cs" />
     <Compile Include="Common\ResumableDownloadManager.cs" />
     <Compile Include="Common\ResumableUploadManager.cs" />
-    <Compile Include="Domain\DownloadFileRequest.cs" />
+    <Compile Include="Domain\DownloadObjectRequest.cs" />
     <Compile Include="Common\Internal\StreamTransferTracker.cs" />
   </ItemGroup>
   <ItemGroup>
