+++ test/TestCase/ObjectTestCase/ObjectBasicOperationTest.cs
@@ -377,226 +377,6 @@ namespace Aliyun.OSS.Test.TestClass.ObjectTestClass
             }
         }
 
-        [Test]
-        public void ResumableUploadObjectTestWithObjectLessThanPartSize()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            try
-            {
-                var fileInfo = new FileInfo(Config.MultiUploadTestFile);
-                var fileSize = fileInfo.Length;
-
-                var result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(), null, fileSize + 1);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectTestWithObjectEqualPartSize()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            try
-            {
-                var fileInfo = new FileInfo(Config.MultiUploadTestFile);
-                var fileSize = fileInfo.Length;
-
-                var result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(), null, fileSize);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectTestWithObjectMoreThanPartSize()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            try
-            {
-                var fileInfo = new FileInfo(Config.MultiUploadTestFile);
-                var fileSize = fileInfo.Length;
-
-                var result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(), Config.DownloadFolder, fileSize - 1);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectTestWithObjectPartSizeTooSmall()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            try
-            {
-                var fileInfo = new FileInfo(Config.MultiUploadTestFile);
-                var fileSize = fileInfo.Length;
-
-                var result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(), Config.DownloadFolder, 1);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectTestWithSmallObjectCheckContentType()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-            var newFileName = Path.GetDirectoryName(Config.UploadTestFile) + "/newfile.js";
-
-            try
-            {
-                File.Copy(Config.UploadTestFile, newFileName);
-
-                var result = _ossClient.ResumableUploadObject(_bucketName, key, newFileName, new ObjectMetadata(), null);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-
-                var objectMeta = _ossClient.GetObjectMetadata(_bucketName, key);
-                Assert.AreEqual("application/x-javascript", objectMeta.ContentType);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-                File.Delete(newFileName);
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectTestWithBigObjectCheckContentType()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-            var newFileName = Path.GetDirectoryName(Config.MultiUploadTestFile) + "/newfile.js";
-
-            try
-            {
-                File.Copy(Config.MultiUploadTestFile, newFileName, true);
-                var fileInfo = new FileInfo(newFileName);
-                var fileSize = fileInfo.Length;
-
-                var result = _ossClient.ResumableUploadObject(_bucketName, key, newFileName, new ObjectMetadata(), Config.DownloadFolder, fileSize / 3);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-
-                var objectMeta = _ossClient.GetObjectMetadata(_bucketName, key);
-                Assert.AreEqual("application/x-javascript", objectMeta.ContentType);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-                File.Delete(newFileName);
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectTestWithBigObject()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-           
-            try
-            {
-                var fileInfo = new FileInfo(Config.MultiUploadTestFile);
-                var fileSize = fileInfo.Length;
-                var config = new ClientConfiguration();
-                config.MaxPartCachingSize = 1024 * 20; // set the max part caching size to 20k so that the ResumableUploadObject will use single thread.
-                var client = OssClientFactory.CreateOssClient(config);
-                var result = client.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(), Config.DownloadFolder, fileSize / 3);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectTestWithManyParts()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            try
-            {
-                var fileInfo = new FileInfo(Config.MultiUploadTestFile);
-                var fileSize = fileInfo.Length;
-                var result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(), Config.DownloadFolder, fileSize / 64);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectTestWithSingleThread()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            try
-            {
-                var fileInfo = new FileInfo(Config.MultiUploadTestFile);
-                var fileSize = fileInfo.Length;
-                var config = new ClientConfiguration();
-                config.MaxResumableUploadThreads = 1;
-                var client = OssClientFactory.CreateOssClient(config);
-                var result = client.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(), Config.DownloadFolder, fileSize / 3);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
         #endregion
 
         #region invalid key
@@ -1802,190 +1582,6 @@ namespace Aliyun.OSS.Test.TestClass.ObjectTestClass
 
         #endregion
 
-        #region resumable put object
-
-        [Test]
-        public void ResumableUploadObjectTest()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            try
-            {
-                var result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, null,
-                                                           Config.DownloadFolder);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectCheckpointTest()
-        {
-            var key = "test/短板.mp4";
-            // this test case requires to run under admin 
-            try
-            {
-                // checkpoint is null
-                var result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, null, Config.DownloadFolder);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-
-                // checkpoint is xx/
-                result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(),
-                                                        Config.DownloadFolder + Path.DirectorySeparatorChar);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-
-                // checkpoint is empty
-                result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, null, "");
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-
-                // checkpoint is current directory
-                result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, null, ".");
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-
-                // checkpoint is previous directory
-                result = _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, null, "..");
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        public class FakeClient : OssClient
-        {
-            public int beginFailedIndex = 0;
-            public int endFailedIndex = 0;
-            public int currentIndex = 0;
-
-            public FakeClient(string endpoint, string accessKeyId, string accessKeySecret)
-            : base(endpoint, accessKeyId, accessKeySecret) { }
-
-            /// <inheritdoc/>        
-            override protected void ThrowIfNullRequest<TRequestType>(TRequestType request)
-            {
-                if (currentIndex >= beginFailedIndex && currentIndex <= endFailedIndex)
-                {
-                    currentIndex++;
-                    throw new ArgumentNullException("uploadPartRequest");
-                }
-
-                currentIndex++;
-                base.ThrowIfNullRequest(request);
-            }
-        };
-
-        [Test]
-        public void ResumableUploadObjectWithRetry()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            try
-            {
-                var client = new FakeClient(Config.Endpoint, Config.AccessKeyId, Config.AccessKeySecret);
-                client.beginFailedIndex = 2;
-                client.endFailedIndex = 3;
-
-                var result = client.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, null,
-                                                        Config.DownloadFolder);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectWithFailedTimeMoreThanRetryTime()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            try
-            {
-                var client = new FakeClient(Config.Endpoint, Config.AccessKeyId, Config.AccessKeySecret);
-                client.beginFailedIndex = 2;
-                client.endFailedIndex = 100;
-                client.currentIndex = 0;
-
-                client.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, null,
-                                                        Config.DownloadFolder);
-                Assert.IsFalse(true);
-            }
-            catch (Exception)
-            {
-                Assert.IsTrue(true);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableUploadObjectFirstFailedAndSecondSucceeded()
-        {
-            var key = OssTestUtils.GetObjectKey(_className);
-
-            var client = new FakeClient(Config.Endpoint, Config.AccessKeyId, Config.AccessKeySecret);
-            try
-            {
-                client.beginFailedIndex = 2;
-                client.endFailedIndex = 100;
-
-                client.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, null,
-                                                       Config.DownloadFolder);
-                Assert.IsTrue(false);
-            }
-            catch (Exception)
-            {
-                Assert.IsTrue(true);
-            }
-
-            try
-            {
-                client.beginFailedIndex = 0;
-                client.endFailedIndex = 0;
-                client.currentIndex = 1;
-
-                var result = client.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, null,
-                                                       Config.DownloadFolder);
-                Assert.IsTrue(OssTestUtils.ObjectExists(_ossClient, _bucketName, key));
-                Assert.IsTrue(result.ETag.Length > 0);
-            }
-            finally
-            {
-                if (OssTestUtils.ObjectExists(_ossClient, _bucketName, key))
-                {
-                    _ossClient.DeleteObject(_bucketName, key);
-                }
-            }
-        }
-
-        #endregion
-
         #region Restore object tests
         [Test]
         public void RestoreObjectBasicTest()
@@ -2034,493 +1630,6 @@ namespace Aliyun.OSS.Test.TestClass.ObjectTestClass
                 {
                     _ossClient.DeleteObject(_archiveBucketName, key);
                 }
-            }
-        }
-        #endregion
-
-        #region Resumable Download objects;
-        [Test]
-        public void ResumableDownloadSmallFileBasicTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, _objectKey, targetFile);
-                var metadata =_ossClient.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadSmallFileProgressUpdateTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-            bool progressUpdateCalled = false;
-            int percentDone = 0;
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, _objectKey, targetFile);
-                long totalBytesDownloaded = 0;
-                request.StreamTransferProgress += (object sender, StreamTransferProgressArgs e) => {
-                    totalBytesDownloaded += e.TransferredBytes;
-                    Console.WriteLine("TotalBytes:" + e.TotalBytes);
-                    Console.WriteLine("TransferredBytes:" + e.TransferredBytes);
-                    Console.WriteLine("PercentageDone:" + e.PercentDone);
-                    progressUpdateCalled = true;
-                    percentDone = e.PercentDone;
-                };
-
-                var metadata = _ossClient.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-                Assert.IsTrue(progressUpdateCalled);
-                Assert.IsTrue(percentDone == 100);
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadSmallFileWithMd5CheckTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-            var config = new ClientConfiguration();
-            config.EnalbeMD5Check = true;
-            var client = OssClientFactory.CreateOssClient(config);
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, _objectKey, targetFile);
-                var metadata = client.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileBasicTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.PutObject(_bucketName, key, Config.MultiUploadTestFile);
-
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                var metadata = _ossClient.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileSingleThreadTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.PutObject(_bucketName, key, Config.MultiUploadTestFile);
-            var config = new ClientConfiguration();
-            config.MaxResumableDownloadThreads = 1;
-            var client = OssClientFactory.CreateOssClient(config);
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                var metadata = client.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileSingleThreadWithMd5CheckTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.PutObject(_bucketName, key, Config.MultiUploadTestFile);
-            var config = new ClientConfiguration();
-            config.MaxResumableDownloadThreads = 1;
-            config.EnalbeMD5Check = true;
-            var client = OssClientFactory.CreateOssClient(config);
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                var metadata = client.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileSingleThreadWithProgressUpdateTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.PutObject(_bucketName, key, Config.MultiUploadTestFile);
-            bool progressUpdateCalled = false;
-            int percentDone = 0;
-            long totalBytesDownloaded = 0;
-            var config = new ClientConfiguration();
-            config.MaxResumableDownloadThreads = 1;
-            config.EnalbeMD5Check = true;
-            var client = OssClientFactory.CreateOssClient(config);
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                request.StreamTransferProgress += (object sender, StreamTransferProgressArgs e) => {
-                    totalBytesDownloaded += e.TransferredBytes;
-                    Console.WriteLine("TotalBytes:" + e.TotalBytes);
-                    Console.WriteLine("TransferredBytes:" + e.TransferredBytes);
-                    Console.WriteLine("PercentageDone:" + e.PercentDone);
-                    progressUpdateCalled = true;
-                    percentDone = e.PercentDone;
-                };
-
-                var metadata = client.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-                Assert.AreEqual(progressUpdateCalled, true);
-                Assert.AreEqual(percentDone, 100);
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileWithProgressUpdateTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.PutObject(_bucketName, key, Config.MultiUploadTestFile);
-            bool progressUpdateCalled = false;
-            int percentDone = 0;
-            long totalBytesDownloaded = 0;
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                request.StreamTransferProgress += (object sender, StreamTransferProgressArgs e) => {
-                    totalBytesDownloaded += e.TransferredBytes;
-                    Console.WriteLine("TotalBytes:" + e.TotalBytes);
-                    Console.WriteLine("TransferredBytes:" + e.TransferredBytes);
-                    Console.WriteLine("PercentageDone:" + e.PercentDone);
-                    progressUpdateCalled = true;
-                    percentDone = e.PercentDone;
-                };
-
-                var metadata = _ossClient.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-                Assert.AreEqual(progressUpdateCalled, true);
-                Assert.AreEqual(percentDone, 100);
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileWithMd5CheckTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.PutObject(_bucketName, key, Config.MultiUploadTestFile);
-
-            var config = new ClientConfiguration();
-            config.EnalbeMD5Check = true;
-            var client = OssClientFactory.CreateOssClient(config);
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                var metadata = client.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileWithMultipartUploadTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(), Config.DownloadFolder, null);
-
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                var metadata = _ossClient.ResumableDownloadFile(request);
-
-                Assert.AreEqual(metadata.ContentLength, new FileInfo(targetFile).Length);
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileWithMultipartUploadWithMd5CheckTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.ResumableUploadObject(_bucketName, key, Config.MultiUploadTestFile, new ObjectMetadata(), Config.DownloadFolder, null);
-
-            try
-            {
-                var config = new ClientConfiguration();
-                config.EnalbeMD5Check = true;
-                var client = OssClientFactory.CreateOssClient(config);
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                var metadata = client.ResumableDownloadFile(request);
-
-                Assert.AreEqual(metadata.ContentLength, new FileInfo(targetFile).Length);
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileBreakAndResumeTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.PutObject(_bucketName, key, Config.MultiUploadTestFile);
-            bool progressUpdateCalled = false;
-            bool faultInjected = false;
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                request.CheckpointFile = Config.DownloadFolder;
-                request.PartSize = 1024 * 1024;
-                request.StreamTransferProgress += (object sender, StreamTransferProgressArgs e) => {
-                    if (!progressUpdateCalled)
-                    {
-                        progressUpdateCalled = true;
-                    }
-                    else
-                    {
-                        if (!faultInjected)
-                        {
-                            faultInjected = true;
-                            throw new TimeoutException("Inject failure");
-                        }
-                    }
-                };
-
-                var metadata = _ossClient.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-                Assert.AreEqual(progressUpdateCalled, true);
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
-            }
-        }
-
-        [Test]
-        public void ResumableDownloadBigFileSingleThreadBreakAndResumeTest()
-        {
-            var targetFile = OssTestUtils.GetTargetFileName(_className);
-            targetFile = Path.Combine(Config.DownloadFolder, targetFile);
-
-            var key = OssTestUtils.GetObjectKey(_className);
-            _ossClient.PutObject(_bucketName, key, Config.MultiUploadTestFile);
-            bool progressUpdateCalled = false;
-            bool faultInjected = false;
-            try
-            {
-                DownloadFileRequest request = new DownloadFileRequest(_bucketName, key, targetFile);
-                request.CheckpointFile = Config.DownloadFolder;
-                request.StreamTransferProgress += (object sender, StreamTransferProgressArgs e) => {
-                    if (!progressUpdateCalled)
-                    {
-                        progressUpdateCalled = true;
-                    }
-                    else
-                    {
-                        if (!faultInjected)
-                        {
-                            faultInjected = true;
-                            throw new TimeoutException("Inject failure");
-                        }
-                    }
-                };
-
-                var config = new ClientConfiguration();
-                config.MaxResumableDownloadThreads = 1;
-                var client = OssClientFactory.CreateOssClient(config);
-
-                var metadata = client.ResumableDownloadFile(request);
-                var expectedETag = metadata.ETag;
-                var downloadedFileETag = FileUtils.ComputeContentMd5(targetFile);
-                Assert.AreEqual(expectedETag.ToLowerInvariant(), downloadedFileETag.ToLowerInvariant());
-                Assert.AreEqual(progressUpdateCalled, true);
-            }
-            finally
-            {
-                try
-                {
-                    FileUtils.DeleteFile(targetFile);
-                }
-                catch (Exception e)
-                {
-                    LogUtility.LogWarning("Delete file {0} failed with Exception {1}", targetFile, e.Message);
-                }
-                _ossClient.DeleteObject(_bucketName, key);
             }
         }
         #endregion
@@ -2627,7 +1736,6 @@ namespace Aliyun.OSS.Test.TestClass.ObjectTestClass
         {
             string symlink = _objectKey + " +\t#_link";
             ObjectMetadata metadata = new ObjectMetadata();
-            //metadata.HttpMetadata.Clear();
             metadata.HttpMetadata.Add("x-oss-header1", "value1");
             metadata.UserMetadata.Add("meta1", "value1");
             metadata.UserMetadata.Add("meta2", "value2");
